app_frontend/requirements_frontend.txt
fastapi==0.88.0
gradio==3.14.0
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.3.0
requests==2.28.1
